requirements.txt
numpy==1.22.4
Pillow==9.5.0
matplotlib==3.7.1
scipy==1.10.1